+++ CHANGELOG.md
@@ -10,9 +10,9 @@ Download this extension from the [VSGallery](https://marketplace.visualstudio.co
 - [x] Table alignment - now aligns all rows in a block.
 - [x] Text is not coloured correctly when a tab is used between keyword and parameter.
 - [x] Removed reference to System in item templates to prevent error when using new project format.
+- [x] Use TextTemplating.VSHost package for SingleFileGenerator.
 - [ ] Support Format Document (Ctrl+K, Ctrl+D)
 - [ ] Support Comments from toolbar (Ctrl+K, Ctrl+C / Ctrl+U)
-- [ ] "Copy Sql To" fails on projects using new format.
 
 Features that have a checkmark are complete and available for
 download in the [CI build](http://vsixgallery.com/author/danny%20quinn).